+++ apps/landing/next.config.ts
@@ -8,7 +8,6 @@ export default withMDX(
     {
       pageExtensions: ['js', 'jsx', 'md', 'mdx', 'ts', 'tsx'],
       output: 'export',
-      assetPrefix: '/devup-ui',
     },
     {},
   ),