+++ apps/landing/src/components/Header/Menu.tsx
@@ -2,8 +2,6 @@
 import { Text } from '@devup-ui/react'
 import { usePathname } from 'next/navigation'
 
-import { URL_PREFIX } from '../../constants'
-
 interface MenuProps {
   children?: React.ReactNode
   keyword: string
@@ -11,9 +9,7 @@ interface MenuProps {
 
 export function Menu({ children, keyword }: MenuProps) {
   const path = usePathname()
-  const selected =
-    path.startsWith(`${URL_PREFIX}/${keyword}`) ||
-    path.startsWith(`/${keyword}`)
+  const selected = path.startsWith(`/${keyword}`)
   return (
     <Text
       _active={{