+++ apps/landing/src/components/Header/index.tsx
@@ -2,7 +2,6 @@ import { Box, css, Flex } from '@devup-ui/react'
 import Link from 'next/link'
 import { Suspense } from 'react'
 
-import { URL_PREFIX } from '../../constants'
 import { Logo } from '../Logo'
 import { Discord } from './Discord'
 import { Github } from './Github'
@@ -50,7 +49,7 @@ export function Header() {
               textDecoration: 'none',
               ml: 4,
             })}
-            href={URL_PREFIX + '/'}
+            href="/"
           >
             <Logo />
           </Link>
@@ -63,7 +62,7 @@ export function Header() {
                   textDecoration: 'none',
                   ml: 4,
                 })}
-                href={URL_PREFIX + '/'}
+                href="/"
               >
                 <Logo />
               </Link>
@@ -75,7 +74,7 @@ export function Header() {
                   textDecoration: 'none',
                   ml: 4,
                 })}
-                href={URL_PREFIX + '/'}
+                href="/"
               >
                 <Logo />
               </Link>
@@ -89,7 +88,7 @@ export function Header() {
             className={css({
               textDecoration: 'none',
             })}
-            href={URL_PREFIX + '/docs/overview'}
+            href="/docs/overview"
           >
             <Menu keyword="docs">Docs</Menu>
           </Link>
@@ -99,7 +98,7 @@ export function Header() {
             className={css({
               textDecoration: 'none',
             })}
-            href={URL_PREFIX + '/team'}
+            href="/team"
           >
             <Menu keyword="team">Team</Menu>
           </Link>