+++ bindings/devup-ui-wasm/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @devup-ui/wasm
 
+## 1.0.2
+
+### Patch Changes
+
+- bb2f49f: Add vendor properties (moz, webkit, ms)
+
+## 1.0.1
+
+### Patch Changes
+
+- 94681ff: to phf
+- 864c37b: Add info to package.json
+
+## 1.0.0
+
+### Major Changes
+
+- 18b1bd7: Optimize Zero value
+
 ## 0.1.58
 
 ### Patch Changes