+++ packages/webpack-plugin/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @devup-ui/webpack-plugin
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [bb2f49f]
+  - @devup-ui/wasm@1.0.2
+
+## 1.0.1
+
+### Patch Changes
+
+- 864c37b: Add info to package.json
+- Updated dependencies [94681ff]
+- Updated dependencies [864c37b]
+  - @devup-ui/wasm@1.0.1
+
+## 1.0.0
+
+### Major Changes
+
+- 18b1bd7: Optimize Zero value
+
+### Patch Changes
+
+- Updated dependencies [18b1bd7]
+  - @devup-ui/wasm@1.0.0
+
 ## 0.1.64
 
 ### Patch Changes